@@ -1,5 +1,6 @@
 {
   "name": "another-to-do",
+  "type": "module",
   "version": "1.0.0",
   "description": "Yet another to do list app",
   "main": "main.js",
@@ -16,28 +17,28 @@
   "author": "Adrian Moldovan",
   "license": "MIT",
   "devDependencies": {
-    "@electron-forge/cli": "^6.0.0-beta.63",
-    "@electron-forge/maker-deb": "^6.0.0-beta.63",
-    "@electron-forge/maker-rpm": "^6.0.0-beta.63",
-    "@electron-forge/maker-squirrel": "^6.0.0-beta.63",
-    "@electron-forge/maker-zip": "^6.0.0-beta.63",
-    "@rollup/plugin-commonjs": "^17.0.0",
-    "@rollup/plugin-node-resolve": "^11.0.0",
-    "concurrently": "^7.0.0",
-    "electron": "^17.1.1",
-    "prettier": "^2.5.1",
-    "prettier-plugin-svelte": "^2.6.0",
+    "@electron-forge/cli": "^7.2.0",
+    "@electron-forge/maker-deb": "^7.2.0",
+    "@electron-forge/maker-rpm": "^7.2.0",
+    "@electron-forge/maker-squirrel": "^7.2.0",
+    "@electron-forge/maker-zip": "^7.2.0",
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "concurrently": "^8.2.2",
+    "electron": "^28.1.4",
+    "prettier": "^3.2.4",
+    "prettier-plugin-svelte": "^3.1.2",
     "rollup": "^2.3.4",
     "rollup-plugin-css-only": "^3.1.0",
     "rollup-plugin-livereload": "^2.0.0",
-    "rollup-plugin-svelte": "^7.0.0",
+    "rollup-plugin-svelte": "^7.1.6",
     "rollup-plugin-terser": "^7.0.0",
-    "svelte": "^3.0.0"
+    "svelte": "^4.2.9"
   },
   "dependencies": {
-    "date-fns": "^2.28.0",
-    "dexie": "^3.2.1",
-    "electron-is-dev": "^2.0.0",
+    "date-fns": "^3.3.1",
+    "dexie": "^3.2.4",
+    "electron-is-dev": "^3.0.1",
     "electron-reload": "^2.0.0-alpha.1",
     "electron-squirrel-startup": "^1.0.0",
     "sirv-cli": "^2.0.0"
